+++ client/src/views/EventModal.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState, VFC } from 'react';
+import { useEffect, useRef, VFC } from 'react';
 import { useParams } from 'react-router-dom';
 import { Dialog, DialogBackdrop, useDialogState } from 'reakit/Dialog';
 
@@ -9,9 +9,6 @@ import { Favourite } from 'components/Favourite';
 export const EventModal: VFC = () => {
   const { slug } = useParams<'slug'>();
   const { event, closeEvent } = useEvent(slug);
-
-  // Controls the favourite icon state
-  const [ isFavourite, setIsFavourite ] = useState(false);
 
   const { currentRef } = useEventContext();
   const { translation } = useGlobalContext();
@@ -26,23 +23,6 @@ export const EventModal: VFC = () => {
       modalRef.current?.focus();
     }
   }, [closeEvent, dialog.visible, dialog.animating]);
-
-  useEffect(() => {
-    const previousFavourites = JSON.parse(localStorage.getItem('WAPPU_FAVOURITES') || '[]');
-    setIsFavourite(previousFavourites.includes(event?.slug));
-  }, [setIsFavourite, event?.slug])
-
-  const addToFavourites = (e: any) => {
-    const previousFavourites = JSON.parse(localStorage.getItem('WAPPU_FAVOURITES') || '[]');
-    let newFavourites = previousFavourites;
-    if (isFavourite) {
-      newFavourites = newFavourites.filter((item: any) => item !== event?.slug);
-    } else {
-      newFavourites = [...previousFavourites, event?.slug];
-    }
-    setIsFavourite(!isFavourite);
-    localStorage.setItem('WAPPU_FAVOURITES', JSON.stringify(newFavourites));
-  }
 
   return (
     <DialogBackdrop
@@ -63,7 +43,6 @@ export const EventModal: VFC = () => {
           crop="1024x512"
           img={event?.content.image}
         />
-        <Favourite fill={isFavourite} addToFavourites={addToFavourites} />
         <div className="flex flex-initial flex-col gap-1 overflow-hidden p-4">
           <h2
             id={labelId}
@@ -72,6 +51,7 @@ export const EventModal: VFC = () => {
             {event?.content.title}
           </h2>
           {event && (
+            <>
             <div className="flex-initial flex-col overflow-auto text-sm">
               <div className="mb-1 flex flex-col gap-1">
                 <EventInfo {...event.content} />
@@ -84,6 +64,11 @@ export const EventModal: VFC = () => {
                 {translation?.backToCalendar}
               </button>
             </div>
+
+            <div className='absolute top-8 right-5'>
+              <Favourite slug={event.slug} size={"30px"} />
+            </div>
+          </>
           )}
         </div>
       </Dialog>
